Data Analysis on Housing and Hotel Bookings
This "Data Analysis Project" involves reading in a house dataset using R and performing various tasks, including data cleaning and transformation. The data is filtered based on specific conditions, and a new variable, `DivSP`, is created by scaling the sale price. Unnecessary columns are removed, and summary statistics such as the mean, standard deviation, and quantiles for variables like "SalePrice" and "LotArea" are computed. A scatter plot is generated to visualize the relationship between "LotArea" and "SalePrice." A multiple linear regression model is built to predict sale prices based on "LotArea" and "FirstFlrSF." Finally, an open-ended analysis is conducted using a hotel bookings dataset, which includes a custom function to summarize both numeric and categorical variables, with options for grouping by certain factors.